pyproject.toml
[project]
name = "searchanddoc-api"
version = "0.1.0"
requires-python = ">=3.9.20"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "certifi==2025.1.31",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "distlib==0.3.9",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "filelock==3.18.0",
    "fixtup==0.1.5",
    "groq>=0.23.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "identify==2.6.9",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "nodeenv==1.9.1",
    "packaging==24.2",
    "platformdirs==4.3.7",
    "pluggy==1.5.0",
    "plumbum==1.9.0",
    "pre-commit==4.2.0",
    "prompt-toolkit==3.0.50",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pysolr==3.10.0",
    "pytest==8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "ray>=2.46.0",
    "requests[security]==2.32.3",
    "rich==13.9.4",
    "rich-toolkit==0.14.0",
    "sentence-transformers>=4.0.2",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "solrq>=1.1.2",
    "starlette==0.46.1",
    "toml==0.10.2",
    "tomli==2.2.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "typer==0.15.2",
    "typing-extensions==4.13.0",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "virtualenv==20.29.3",
    "watchfiles==1.0.4",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
]

[tools.fixtup]
fixtures = "./tests/fixtures"
plugins = [
    "fixtup.plugins.dotenv",
    "fixtup.plugins.docker"
]
